Add some typecasts to match printf format
git-svn-id: https://abacuscm.svn.sourceforge.net/svnroot/abacuscm/trunk@353 ac3b2d11-31fc-0310-b7bc-aecf6fe4fc4f

+++ src/mainwindow.C
@@ -300,9 +300,9 @@ void MainWindow::doTimer() {
 
 		char buffer[30];
 		sprintf(buffer, "%02ld:%02ld:%02ld",
-			remain / 3600,
-			(remain / 60) % 60,
-			remain % 60);
+			(long) remain / 3600,
+			(long) (remain / 60) % 60,
+			(long) remain % 60);
 		clock->setText(buffer);
 	}
 }
